FB/backend/requirements.txt
fastapi
uvicorn[standard]
mediapipe
opencv-python
numpy
pillow
pydantic>=2.0
pydantic-settings>=2.0
